fix: move generateTableHeaders to resource
